tests: move test to factories.spec
